// types/dto.ts
export interface PublisherDetail {
  id: number;
  name: string;
  engName: string;
  logo?: string;
  isOfficial: boolean;
  description?: string;
  urls?: UrlInfo[];
  books: BookDto[];
  tagList?: TagDto;
}

/** @default [] */
export interface UrlInfo {
  url: string;
  type: "Link" | "Youtube" | "Profile" | "Homepage" | "Blog";
}

export interface UrlInfoTypeString {
  url: string;
  type: string;
}

export interface BookDto {
  isbn: string;
  title: string;
  summary: string;
  publisher: Simple;
  /** @format date-time */
  publishedDate: string;
  authorList: Simple[];
  titleImage?: string;
  translator?: string;
  /** @format int32 */
  price?: number;
}

export interface Detail {
  /** @format int64 */
  isbn: string;
  title: string;
  detailUrl: string;
  summary: string;
  publishedDate: Date;
  titleImage: string;
  authorList: [];
  translator: [];
  price: number;
  status: string;
  publisher: {
    id: number;
    name: string;
  };
  contentsDtoList: [];
  eventDtoList: [];
}

export interface Simple {
  /** @format int64 */
  id: number;
  name: string;
}

export interface TagDto {
  name: string;
}
// ...
export interface EventDto {
  /** @format int64 */
  id: number;
  title: string;
  host: string;
  creator: UserDto;
  urls: UrlInfo[];
  bookTitle?: string;
  senderName?: string;
  senderEmail?: string;
  senderMessage?: string;
  location: "ONLINE" | "OFFLINE";
  /** @format date */
  startDate: string;
  /** @format date */
  endDate: string;
  bookIsbnList?: { value: number }[];
  eventType: string;
  eventFlag: "SOLO" | "GROUP" | "ETC";
  memo?: string;
  books: BookDto[];
  tags: TagDto[];
}

export interface UserDto {
  /** @format int64 */
  id: number;
  name: string;
  bookIsbnList?: number[];
}
// ...
export interface ContentsDto {
  /** @format int64 */
  id: number;
  urls: UrlInfo[];
  title?: string;
  image?: string;
  creator: UserDto;
}

export interface ContactDto {
  id: number;
  email: string;
  message: string;
  name?: string;
}
// ...
export interface AuthorDto {
  /** @format int64 */
  id: number;
  name: string;
  description?: string;
  profile?: string;
  bookList: BookDto[];
}

export interface PageResponsePublisherDto {
  items: PublisherDto[];
  /** @format int64 */
  totalCount: number;
  /** @format int32 */
  totalPages: number;
  hasNext: boolean;
  hasPrevious: boolean;
}

export interface PublisherDto {
  /** @format int64 */
  id: number;
  name: string;
  engName?: string;
  logo?: string;
  isOfficial: boolean;
  description?: string;
  urls: UrlInfo[];
}

export interface PageResponseEventDto {
  items: EventDto[];
  /** @format int64 */
  totalCount: number;
  /** @format int32 */
  totalPages: number;
  hasNext: boolean;
  hasPrevious: boolean;
}

export interface ListResponseBookDto {
  items: BookDto[];
  /** @format int32 */
  length: number;
}

export interface PageResponseContentsDto {
  items: ContentsDtoChanged[];
  /** @format int64 */
  totalCount: number;
  /** @format int32 */
  totalPages: number;
  hasNext: boolean;
  hasPrevious: boolean;
}

export interface ListResponseContentsDto {
  items: ContentsDto[];
  /** @format int32 */
  length: number;
}

export interface ListResponseContactDto {
  items: ContactDto[];
  /** @format int32 */
  length: number;
}

export interface PageResponseBookDto {
  items: BookDto[];
  /** @format int64 */
  totalCount: number;
  /** @format int32 */
  totalPages: number;
  hasNext: boolean;
  hasPrevious: boolean;
}

export interface ListResponseEventDto {
  items: EventDto[];
  /** @format int32 */
  length: number;
}

export interface ListResponseRecommendedBookDto {
  items: RecommendedBookDto[];
  /** @format int32 */
  length: number;
}

export interface RecommendedBookDto {
  isbn: string;
  title: string;
  summary: string;
  /** @format date-time */
  publishedDate: string;
  titleImage?: string;
  authorList: AuthorDto[];
  translator?: string;
  /** @format int32 */
  price?: number;
  publisher: PublisherDto;
  /** @format date */
  recommendedDate: string;
}

export interface PageResponseAuthorDto {
  items: AuthorDto[];
  /** @format int64 */
  totalCount: number;
  /** @format int32 */
  totalPages: number;
  hasNext: boolean;
  hasPrevious: boolean;
}
// ...
export interface ContentsDetail {
  id: number;
  title: string;
  urls: UrlInfo;
  image: string;
  booksDtoList: BookDto[];
  tagDtoList: TagDto[];
  creator: UserDto;
}

export interface ContentsDtoChanged {
  /** @format int64 */
  id: number;
  urls: UrlInfo[];
  title?: string;
  image?: string;
  creator: UserDto;
  books: BookDto[];
  tags: TagDto[];
}